requirements/base.txt
Django==3.2
djangorestframework==3.11
psycopg2-binary
Pillow>=5.3.0,<=5.4.0
mysqlclient
django-cors-headers
djangorestframework_simplejwt
django-guardian
djangorestframework-bulk
djangorestframework-guardian
hurry.filesize
django-templated-email
django-split-settings
django-rest-passwordreset
rest-framework-generic-relations
xhtml2pdf
fpdf
PyPDF2
reportlab
convertapi
PyPDF3
pdfrw
pdfkit
drf-yasg
drf-yasg[validation]
imgkit
six